+++ pkg/apis/application/v1alpha1/app_types_test.go
@@ -9,6 +9,7 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/yaml"
 )
@@ -62,7 +63,13 @@ func Test_GenerateAppYAML(t *testing.T) {
 func newAppExampleCR() *App {
 	cr := NewAppCR()
 
-	cr.Name = "prometheus"
+	cr.ObjectMeta = metav1.ObjectMeta{
+		Name:      "prometheus",
+		Namespace: "default",
+		Labels: map[string]string{
+			"app-operator.giantswarm.io/version": "1.0.0",
+		},
+	}
 	cr.Spec = AppSpec{
 		Name:      "prometheus",
 		Namespace: "monitoring",